pyproject.toml
[tool.poetry]
name = "coursework8"
version = "0.1.0"
description = ""
authors = ["Reechka <R2echka@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.1.7"
django-rest-framework = "^0.1.0"
python-dotenv = "^1.0.1"
drf-yasg = "^1.21.10"
celery = "^5.4.0"
django-celery-beat = "^2.7.0"
redis = "^5.2.1"
eventlet = "^0.39.1"
djangorestframework-simplejwt = "^5.5.0"
django-cors-headers = "^4.7.0"
psycopg2 = "^2.9.10"
requests = "^2.32.3"


[tool.poetry.group.develop.dependencies]
coverage = "^7.7.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
